moved mock readable/writable streams used in tests to own module
